Add logo for extension
This logo is based on the file "AWS Simple Icons AWS Cloud.svg"
retrieved from
https://commons.wikimedia.org/wiki/File:AWS_Simple_Icons_AWS_Cloud.svg,
whose author is Amazon Web Services LLC. According to the license the
image is in the public domain.

The addition of the magnifying glass with the accompanying zoomed effect
has been applied by me afterwards, and I am putting the modified image
into the public domain, too.