Change default search behaviours
Instead of searching the (for most of the user base probably more
obscure) CloudFormation documentation, this commit changes that a
non-prefixed query will target the API documentation instead, and that
the `!` prefix must be used to search the CloudFormation docs.

(Personally I would have *loved* to use `:` as the prefix for
CloudFormation, but unfortunately the search-extension-core used
hard-binds `:` to commands. I'm still on the fence about potentially
upstreaming support to allow changing this to a different prefix; I will
make a decision on this before publishing a v1.0.0).