Change the trigger keyword from `aws` to `ase`
After a few weeks of use now I realized that I regularly search for `aws
...` terms in the Omnibox where I intended to actually perform a search
engine search, instead of a search using AWS search extension.
Unfortunately the space character does trigger the keyword and thus drop
the user into a search in AWS search extension, even if they maybe
didn't want it.

To avoid this annoying clash, this commit changes the trigger keyword
from `aws` to `ase` (short for AWS search extension). While not quite as
intuitive as `aws`, it still shares most of the traits that `aws` had as
a keyword:

* It is short with only 3 characters in length.

* It is as easy to type: the characters are all in the left cluster for
  QWERTY, QWERTZ, and AZERTY keyboard layouts, and they are in the home
  row for Dvorak.