Make `UpdateCommand::scheduleIndexUpdates` async
The call-sites of this function do not have to change, since the update
itself ran asynchronous previously as well. The only difference now is
that a call-site has the option to wait if they want to (since we no
longer hide the `Promise`).

The function now also returns the `lastUpdate` timestamp to the caller
(if they await it).