feat(cfn): open the correct URL for SAM docs
The CloudFormation and SAM-docs have different base-URLs where to open
their docs in. To allow the extension to map to the correct base-URL,
we'll have to keep track of the source.

Since this field is added at the end of the index, this change is
backwards-compatible and thus doesn't break this extension for users who
might have an outdated extension-version but a newer index.

Fixes #123.