Add a new version of the CFN index
With the index datasources changing and containing full documentation URLs, it
made sense to no longer build up the URL to the CFN-page in the extension, but
have it in the index instead. We also can and should no longer rely on
documentation to be present, so the newer version of the index makes the
presence of documentation optional.

The change made here does _not_ remove the "old" V1 index, at least not from
the `indices` branch. It does not bundle both versions with the extension,
since that is unnecessary, but old versions of the extension will still be able
to get the version 1 index without issues going forward.